remove auth from explore page routes
