+++ src/app/api/companies/get-company-by-id/route.ts
@@ -1,15 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import supabase from "@/lib/supabase";
-import { PrivyClient } from "@privy-io/server-auth";
-import User from "@/types/user";
-
-const privyClient = new PrivyClient(
-  process.env.NEXT_PUBLIC_PRIVY_APP_ID!,
-  process.env.PRIVY_SECRET!
-);
 
 export async function GET(req: NextRequest, res: NextResponse) {
-  const authToken = req.headers.get("Authorization")?.replace("Bearer ", "");
   const url = new URL(req.url);
   const companyId = url.searchParams.get("companyId");
 
@@ -21,26 +13,6 @@ export async function GET(req: NextRequest, res: NextResponse) {
   }
 
   try {
-    const verificationClaim = await privyClient.verifyAuthToken(authToken!);
-
-    // Get user and check if they are an admin
-    const { data: userData, error: userError } = await supabase
-      .from("users")
-      .select()
-      .eq("privy_did", verificationClaim.userId)
-      .returns<User[]>();
-
-    if (userError) {
-      throw new Error(
-        `Error fetching user for the supplied auth token: ${userError}`
-      );
-    }
-
-    const user = userData?.[0];
-    if (!user.is_admin) {
-      throw new Error("User is not an admin");
-    }
-
     // Get company information
     const { data: companyData, error: companyError } = await supabase
       .from("companies")
